requirements.txt
torch>=2.1.0
transformers>=4.40.0
huggingface_hub>=0.23.0
hf_xet>=0.1.0
numpy>=1.24.0
safetensors>=0.4.0
tqdm>=4.65.0
psutil>=5.9.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.120.0
requests>=2.31.0
click>=8.1.0
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.5.0
